Basic Auth Added and UserDAo refactor
