+++ project-api/gradle/libs.versions.toml
@@ -28,6 +28,7 @@ postgresql = {module = "org.postgresql:postgresql" , version.ref ="postgres" }
 h2 = { module = "com.h2database:h2", version.ref = "h2-version" }
 content-negotiation = { module = "io.ktor:ktor-server-content-negotiation", version.ref = "ktor" }
 ktor-serialization-kotlinx-json = { module = "io.ktor:ktor-serialization-kotlinx-json", version.ref = "ktor" }
+ktor-auth = { module = "io.ktor:ktor-server-auth", version.ref = "ktor"}
 ktor-server-auth = { module = "io.ktor:ktor-server-auth", version.ref = "ktor" }
 koin-ktor = { module = "io.insert-koin:koin-ktor", version.ref = "koin-version" }
 koin-logger-slf4j = { module = "io.insert-koin:koin-logger-slf4j", version.ref = "koin-version" }
