+++ project-api/src/main/kotlin/model/user/DBUserRepository.kt
@@ -1,16 +1,16 @@
 package mx.unam.fciencias.ids.eq1.model.user
 
-import mx.unam.fciencias.ids.eq1.db.UserDAO
+import mx.unam.fciencias.ids.eq1.db.user.UserDAO
 import mx.unam.fciencias.ids.eq1.db.UserTable
+import mx.unam.fciencias.ids.eq1.db.user.UserDAO.Companion.userDaoToModel
 import mx.unam.fciencias.ids.eq1.db.utils.suspendTransaction
-import mx.unam.fciencias.ids.eq1.db.userDaoToModel
 import org.jetbrains.exposed.sql.*
 import org.jetbrains.exposed.sql.transactions.transaction
 import org.koin.core.annotation.Single
 
 
 @Single
-class PostgresqlUserRepository(private val database: Database) : UserRepository {
+class DBUserRepository(database: Database) : UserRepository {
 
     init {
         transaction(database) {
